Login and register update.
Update of the login and register page.
