--- src/App.css
@@ -1,179 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-/* login box form */
-
-
-.frm1{
-  background-color: whitesmoke;
-  margin: 10px;
-  padding: 30px;
-  border-radius: 10px;
-  border: 1px solid rgb(197, 190, 190);
-  /* position: ; */
-}
-
-.form-c{
- 
-  background-color: whitesmoke;
-  /* background-color: slateblue; */
-  margin-bottom:2px ;
-}
-.form-row{
-  padding:0px 30px 0px 30px;
-  color: black;
-  background-color: white;
-  border-radius: 15px;
-  margin: 10px 0px 10px 0px ;
-  border: 1px solid #dadada;
-  
-}
-.lgn.details{
-  background-color: white;
-  /* background-color: thistle; */
-  margin-top: 70px;
-  padding: 15px 15px  10px 15px;
-  border: radius 15px;
-}
-.lgn-inner{
-  background-color: white;
-  /* background-color: #f02199; */
-  display: flex;
-  justify-content: center;
-  /* position: absolute; */
-}
-
-.form-row.btn{
-  margin-bottom: 30px;
-}
-.form-row input{
-  border-bottom: 1px solid white !important;
-  /* color: white !important; */
-}
-.form-row input:active{
-  border-bottom: 1px solid white !important;
-  /* color: white !important; */
-}  
-
-.lgn-row-2{
-  margin-top: 20px;
-  border-bottom: 1px solid #D7D7D7;
-  border-top: 1px solid #A1A1A1;
-  line-height: 0;
-  text-align: center;
-}
-.sid span{
-  font-size: medium;
-   background-color: whitesmoke;
-   padding: 1px 3px 1px 3px ;
-}
-.lgn-row-2{
-  background-color: whitesmoke;
-  margin-bottom: 15px;
-}
-
-.blur{
-  -webkit-filter: blur(4px);
--moz-filter: blur(4px);
--ms-filter: blur(4px);
--o-filter: blur(4px);
-filter: blur(4px);
-}
-/* .app-logos{
-  background-color: #b30202;
-  padding: 20px;
-} */
-.app-lg{
-  margin-left: 25px;
-  padding:10px 20px 10px 10px;
-  border:1px solid black;
-  background-color:whitesmoke;
-  
-}
-.app-lg.a{
-  margin-right: 10px;
-  float:left;
-}
-.app-lg.b{
- float:right;
-}
-
-.regis{
-  margin-top: 25px;
-  padding: 10px;
-  float: right;
-}
-.fano-text{
-  display: none;
-}
-@media all and (max-width:500px ){
-  .fano-text{
-    display:contents !important;
-  }
-  .frm1{
-    background-color: whitesmoke;
-    margin: 10px;
-    padding: 30px;
-    border: none;
-    /* border-top: 10px; */
-    /* position: ; */
-  }
-  h1.fano-text{
-    font-weight: 800; 
-    color: #61dafb;
-  }
-  .lgn.details{
-    background-color: whitesmoke;
-    border-top: 1px solid rgb(163, 151, 151);
-    border-top-left-radius: 50px;
-    border-top-right-radius: 50px;
-  }
-  .form-c{
-    background-color: #61dafb;
-  }
-  .lgn-inner{
-    background-color: whitesmoke;
-  }
-
-
-
-}
-
-
-
