--- src/components/Footer.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const Footer = () => {
-    return (
-        <div>
-            <footer className="page-footer footer-copyright blue lighten-3" >
-                <div className="container">
-                © 2021 Fano
-            </div>
-            </footer>
-        </div>
-    )
-}
-
-export default Footer
