--- src/index.js
@@ -1,63 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import {createStore, applyMiddleware, compose} from 'redux'
-import rootReducer from './store/reducers/rootReducer'
-import {Provider, useSelector} from 'react-redux'
-import thunk from 'redux-thunk'
-import { createFirestoreInstance, reduxFirestore, getFirestore } from 'redux-firestore'
-import { ReactReduxFirebaseProvider, getFirebase, isLoaded } from 'react-redux-firebase'
-import firebase from './config/fbconfig'
-import 'firebase/firestore'
-
-const rrfConfig = { 
-  userProfile: 'users',
-  presence: 'presence',
-  sessions:'sessions',
-  useFirestoreForProfile: true
-}
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-const store = createStore(rootReducer, 
-  composeEnhancers(
-    applyMiddleware(thunk.withExtraArgument({getFirebase, getFirestore})),
-    reduxFirestore(firebase),
-    )
-  );
-
-const rffProps = {
-  firebase,
-  useFirestoreForProfile: true,
-  config: rrfConfig,
-  dispatch: store.dispatch,
-  createFirestoreInstance
-}
-
-function AuthIsLoaded({children}) {
-  const auth = useSelector(state => state.firebase.auth)
-  if (!isLoaded(auth)) return <div>Loading Screen...</div>;
-
-  return children
-
-}
-
-
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <ReactReduxFirebaseProvider {...rffProps}>
-        <AuthIsLoaded>
-          <App />
-        </AuthIsLoaded>
-      </ReactReduxFirebaseProvider>
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
