--- src/store/actions/authActions.js
@@ -1,44 +1,0 @@
-import { firestoreReducer } from "redux-firestore";
-
-export const signIn = (credentials) => {
-    return (dispatch, getState, { getFirebase, getFirestore }) => {
-
-        
-        firebase.auth().signInWithEmailAndPassword(
-            credentials.email,
-            credentials.password
-        ).then(() => {
-            dispatch({ type: "LOGIN_SUCCESS" })
-        }).catch((err) => {
-            dispatch({ type: "LOGIN_ERROR", err })
-        })
-
-    }
-}
-
-export const signOut = () => {
-    return (dispatch, getState, { getFirebase, getFirestore }) => {
-
-        const firebase = getFirebase();
-        firebase.auth().signOut()
-        .then(() => {
-            dispatch({ type: "SIGNOUT_SUCCESS" })
-        }).catch((err) => {
-            dispatch({ type: "SIGNOUT_ERROR", err })
-        })
-    }
-}
-
-// TODO signUp function
-export const signUp = (email, password, firstName, lastName) => {
-    return (dispatch, getState, { getFirebase, getFirestore }) => {
-        
-    }
-}
-
-// TODO makeAdmin function
-
-// TODO removeAdmin
-
-
-
