--- src/store/actions/imageActions.js
@@ -1,6 +1,0 @@
-// TODO fetchNextImage
-
-// TODO newAnnotation
-
-// TODO uploadNewImages
-
