--- src/store/reducers/rootReducer.js
@@ -1,13 +1,0 @@
-import { firebaseReducer } from 'react-redux-firebase'
-import {firestoreReducer} from 'redux-firestore'
-import {authReducer} from './authReducer'
-
-import {combineReducers} from 'redux'
-
-const rootReducer = combineReducers({
-    auth: authReducer,
-    firebase: firebaseReducer,
-    firestore: firestoreReducer
-})
-
-export default rootReducer
